refactor: extract tag parsing to webview module & restore .prettierrc
Address PR #174 review feedback:
- Extract tag parsing into webview/utils/tags.ts, remove tagMap prop
- Restore .prettierrc, remove inline prettier config from package.json
- Remove unused tagsData from shared.ts
- Remove extensionKind restriction from package.json
